once more with feeling
Fixing CI to properly report test coverage and to actually execute the compile tests.
