+++ .github/workflows/cetlvast.yml
@@ -15,14 +15,22 @@ jobs:
     container: ghcr.io/opencyphal/toolshed:ts22.4.3
     strategy:
       matrix:
-        std: [base, intermediate, target]
+        build_flavor: [Debug, Release]
+        exceptions: [exceptions, no-exceptions]
+        std: [14, 17, 20]
         toolchain: [gcc, clang]
+        include:
+          - build_flavor: Coverage
+            exceptions: no-exceptions
+            std: 14
+            toolchain: gcc
+        exclude:
+          - build_flavor: Debug
+            exceptions: no-exceptions
+            std: 14
+            toolchain: gcc
     steps:
     - uses: actions/checkout@v3
-    - name: setup environment common
-      run: >
-        echo "CETLVAST_EXT_PATH=./cetlvast/build_external" >> $GITHUB_ENV;
-        echo "CETLVAST_SUITE_BUILD_PATH=./build" >> $GITHUB_ENV;
 
     - name: Cache ext modules
       id: cetlvast-ext
@@ -30,7 +38,7 @@ jobs:
       env:
         cache-name: cetlvast-ext-cache
       with:
-        path: ${{ env.CETLVAST_EXT_PATH }}
+        path: cetlvast/build_external
         key: gtest_b796f7d44681514f58a683a3a71ff17c94edb0c1
     - name: run tests
       run: >
@@ -38,35 +46,36 @@ jobs:
         -v
         --online
         --asserts
-        --build-flavor Coverage
+        --${{ matrix.exceptions }}
+        --build-flavor ${{ matrix.build_flavor }}
         --cpp-standard ${{ matrix.std }}
         --toolchain ${{ matrix.toolchain }}
         test
+    - name: debug output
+      if: ${{ runner.debug == '1' }}
+      run: ls -lAhR build/
     - name: upload-artifacts
       uses: actions/upload-artifact@v3
       with:
-        name: release-${{ matrix.std }}-${{ matrix.toolchain }}
+        name: ${{ matrix.build_flavor }}-${{ matrix.exceptions }}-${{ matrix.std }}-${{ matrix.toolchain }}
         path: |
-          ${{ env.CETLVAST_SUITE_BUILD_PATH }}/compile_commands.json
-          ${{ env.CETLVAST_SUITE_BUILD_PATH }}/cetlvast/suites/unittest
-          ${{ env.CETLVAST_SUITE_BUILD_PATH }}/cetlvast/suites/compile
+          build/compile_commands.json
+          build/*/**/coverage.xml
+          build/*/**/*-sonarqube.xml
+          build/*/**/gcovr_html/*.*
         if-no-files-found: error
   docs:
     runs-on: ubuntu-latest
     container: ghcr.io/opencyphal/toolshed:ts22.4.3
     steps:
     - uses: actions/checkout@v3
-    - name: setup environment
-      run: >
-        echo "CETLVAST_EXT_PATH=./cetlvast/build_external" >> $GITHUB_ENV;
-        echo "CETLVAST_SUITE_BUILD_PATH=./build" >> $GITHUB_ENV;
     - name: Cache ext modules
       id: cetlvast-ext
       uses: actions/cache@v3
       env:
         cache-name: cetlvast-ext-cache
       with:
-        path: ${{ env.CETLVAST_EXT_PATH }}
+        path: cetlvast/build_external
         key: gtest_b796f7d44681514f58a683a3a71ff17c94edb0c1
     - name: doc-gen
       run: >
@@ -78,13 +87,20 @@ jobs:
         --cpp-standard 14
         build-docs
     - name: Setup Pages
+      if: ${{ github.event_name != 'pull_request' }}
       uses: actions/configure-pages@v3
     - name: Upload docs
       if: ${{ github.event_name != 'pull_request' }}
       uses: actions/upload-pages-artifact@v1
       with:
-        path: "${{ env.CETLVAST_SUITE_BUILD_PATH }}/suits/docs/html/"
-
+        path: "build/suites/docs/html/"
+    - name: upload-pr-docs
+      if: ${{ github.event_name == 'pull_request' }}
+      uses: actions/upload-artifact@v3
+      with:
+        name: pr-docs
+        path: "build/suites/docs/html/"
+        if-no-files-found: error
   sonarqube:
     needs:
     - verification
@@ -97,10 +113,6 @@ jobs:
           fetch-depth: 0  # Shallow clones should be disabled for a better relevancy of analysis by sonarqube.
     - name: Install sonar-scanner
       uses: SonarSource/sonarcloud-github-c-cpp@v1
-    - name: setup environment
-      run: >
-        echo "CETLVAST_EXT_PATH=./cetlvast/build_external" >> $GITHUB_ENV;
-        echo "CETLVAST_SUITE_BUILD_PATH=./build" >> $GITHUB_ENV;
 
     - name: Cache ext modules
       id: cetlvast-ext
@@ -108,14 +120,12 @@ jobs:
       env:
         cache-name: cetlvast-ext-cache
       with:
-        path: ${{ env.CETLVAST_EXT_PATH }}
+        path: cetlvast/build_external
         key: gtest_b796f7d44681514f58a683a3a71ff17c94edb0c1
-
     - uses: actions/download-artifact@v3
       with:
-        name: unittest-base-clang
-        path: ${{ env.CETLVAST_SUITE_BUILD_PATH }}
-
+        name: Coverage-no-exceptions-14-gcc
+        path: build
     - name: Run sonar-scanner
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -126,16 +136,18 @@ jobs:
           --define sonar.projectKey=OpenCyphal_CETL
           --define sonar.projectName=CETLVaSt
           --define sonar.projectVersion=1.0
-          --define sonar.sources=include,cetlvast/suites/unittest
+          --define sonar.sources=include,cetlvast/suites/unittest/sonar.cpp
+          --define sonar.tests=cetlvast/suites/unittest,cetlvast/suites/docs/examples
+          --define sonar.test.inclusions=test_*.cpp,example_*.cpp
           --define sonar.sourceEncoding=UTF-8
           --define sonar.host.url=https://sonarcloud.io
           --define sonar.cfamily.ignoreHeaderComments=false
-          --define sonar.coverage.exclusions="cetlvast/suites/unittest/**/*"
-          --define sonar.cpd.exclusions="cetlvast/suites/unittest/**/*"
-          --define sonar.cfamily.compile-commands=${{ env.CETLVAST_SUITE_BUILD_PATH }}/compile_commands.json
-          --define sonar.cfamily.reportingCppStandardOverride=14
-          --define sonar.coverageReportPaths=${{ env.CETLVAST_COVERAGE_REPORT }}/todo.xml
-          --define sonar.testExecutionReportPaths=${{ env.CETLVAST_SUITE_BUILD_PATH }}/todo.xml
+          --define sonar.coverage.exclusions="cetlvast/suites/unittest/**/*,cetlvast/suites/docs/examples/**/*,**/sonar.cpp"
+          --define sonar.cpd.exclusions="cetlvast/suites/unittest/**/*,cetlvast/suites/docs/examples/**/*,**/sonar.cpp"
+          --define sonar.cfamily.compile-commands="build/compile_commands.json"
+          --define sonar.cfamily.reportingCppStandardOverride=c++14
+          --define sonar.coverageReportPaths="build/suites/unittest/coverage.xml,build/suites/docs/examples/coverage.xml"
+          --define sonar.testExecutionReportPaths="build/suites/unittest/unittest-sonarqube.xml,build/suites/docs/examples/examples-sonarqube.xml"
 
   deploy-docs:
     if: ${{ (github.event_name == 'release' && !github.event.release.prerelease) || github.event_name == 'workflow_dispatch' }}
