+++ .vscode/cmake-variants.json
@@ -5,39 +5,26 @@
         "choices": {
             "debug": {
                 "short": "Debug",
-                "long": "Debug builds, no optimizations.",
-                "buildType": "Debug"
+                "long": "Debug builds with asserts enabled, no optimizations.",
+                "buildType": "Debug",
+                "settings": {
+                    "CETL_ENABLE_DEBUG_ASSERT": true,
+                    "LIBCXX_ENABLE_ASSERTIONS": true
+                }
             },
             "coverage": {
                 "short": "Coverage",
-                "long": "Coverage builds, no optimizations.",
-                "buildType": "Coverage"
+                "long": "Coverage builds with asserts enabled, no optimizations.",
+                "buildType": "Coverage",
+                "settings": {
+                    "CETL_ENABLE_DEBUG_ASSERT": true,
+                    "LIBCXX_ENABLE_ASSERTIONS": true
+                }
             },
             "release": {
                 "short": "Release",
                 "long": "Release builds, some optimizations",
                 "buildType": "Release"
-            }
-        }
-    },
-    "debugAssert":{
-        "default": "off",
-        "choices": {
-            "off": {
-                "short": "no-asserts",
-                "long": "disable intrinsic asserts intended to be used only for testing or verification.",
-                "settings": {
-                    "CETL_ENABLE_DEBUG_ASSERT": false,
-                    "LIBCXX_ENABLE_ASSERTIONS": false
-                }
-            },
-            "on": {
-                "short": "asserts",
-                "long": "enable intrinsic asserts intended to be used only for testing or verification.",
-                "settings": {
-                    "CETL_ENABLE_DEBUG_ASSERT": true,
-                    "LIBCXX_ENABLE_ASSERTIONS": true
-                }
             }
         }
     },
