+++ cetlvast/cmake/modules/Findclangformat.cmake
@@ -10,21 +10,10 @@
 
 find_program(CLANG_FORMAT clang-format)
 
-set(LOCAL_BIN_MODULE_PATHS ${CMAKE_MODULE_PATH})
-list(TRANSFORM LOCAL_BIN_MODULE_PATHS APPEND "/../bin")
-
-find_file(CLANG_FORMAT_PYTHON_SHIM clang-format-check.py
-            PATHS ${LOCAL_BIN_MODULE_PATHS}
-)
-
-if (NOT CLANG_FORMAT_PYTHON_SHIM STREQUAL "CLANG_FORMAT_PYTHON_SHIM-NOTFOUND")
-    set(CLANG_FORMAT_PYTHON_SHIM_FOUND TRUE)
-endif()
-
 include(FindPackageHandleStandardArgs)
 
 find_package_handle_standard_args(clangformat
-    REQUIRED_VARS CLANG_FORMAT CLANG_FORMAT_PYTHON_SHIM_FOUND
+    REQUIRED_VARS CLANG_FORMAT
 )
 
 # +---------------------------------------------------------------------------+
@@ -38,66 +27,62 @@ find_package_handle_standard_args(clangformat
 #                                     ${CMAKE_CURRENT_SOURCE_DIR}
 # :param GLOB_PATTERN glob          - A pattern to match files against.
 # :option ADD_TO_ALL                - If set the target is added to the default build target.
+# :option FORMAT_IN_PLACE           - If set the target is defined as an in-place formatter.
+# :return OUT_TARGET_NAME           - The name of a variable, in the parent scope, to set to the
+#                                     name of the target created.
 #
 function(enable_clang_format_check_for_directory)
     #+-[input]----------------------------------------------------------------+
-    set(options ADD_TO_ALL)
-    set(singleValueArgs GLOB_PATTERN DIRECTORY)
+    set(options ADD_TO_ALL FORMAT_IN_PLACE)
+    set(singleValueArgs GLOB_PATTERN DIRECTORY OUT_TARGET_NAME)
     set(multiValueArgs "")
     cmake_parse_arguments(PARSE_ARGV 0 ARG "${options}" "${singleValueArgs}" "${multiValueArgs}")
 
     if(NOT ARG_DIRECTORY)
         set(ARG_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
     endif()
+
     #+-[body]-----------------------------------------------------------------+
     cmake_path(GET CMAKE_CURRENT_SOURCE_DIR STEM LOCAL_DIRECTORY_NAME)
-    set(LOCAL_TARGET_NAME "${LOCAL_DIRECTORY_NAME}_clang_format_check")
+
+    cmake_path(APPEND ARG_DIRECTORY "${ARG_GLOB_PATTERN}" OUTPUT_VARIABLE LOCAL_GLOB_PATTERN_WITH_PATH)
 
     if  (ARG_ADD_TO_ALL)
-        set(LOCAL_ALL "ALL")
+    set(LOCAL_ALL "ALL")
     else()
-        set(LOCAL_ALL  "")
+    set(LOCAL_ALL  "")
+    endif()
+
+    file(GLOB_RECURSE LOCAL_SOURCE_FILES
+         CONFIGURE_DEPENDS
+         LIST_DIRECTORIES false
+         RELATIVE ${ARG_DIRECTORY}
+         ${LOCAL_GLOB_PATTERN_WITH_PATH}
+    )
+
+    if (ARG_FORMAT_IN_PLACE)
+        set(LOCAL_FORMAT_IN_PLACE "-i")
+        set(LOCAL_TARGET_NAME "danger-danger-${LOCAL_DIRECTORY_NAME}-clang-format-in-place")
+    else()
+        set(LOCAL_FORMAT_IN_PLACE "--dry-run")
+        set(LOCAL_TARGET_NAME "${LOCAL_DIRECTORY_NAME}-clang-format-check")
     endif()
 
     add_custom_target(${LOCAL_TARGET_NAME} ${LOCAL_ALL}
-                      COMMAND ${CLANG_FORMAT_PYTHON_SHIM}
-                              --clang-format-path ${CLANG_FORMAT}
-                              ${ARG_GLOB_PATTERN}
+                      COMMAND ${CLANG_FORMAT}
+                              --fallback-style=none
+                              -style=file
+                              --verbose
+                              --Werror
+                              ${LOCAL_FORMAT_IN_PLACE}
+                              ${LOCAL_SOURCE_FILES}
                       VERBATIM
                       WORKING_DIRECTORY ${ARG_DIRECTORY}
     )
+
+    #+-[output]---------------------------------------------------------------+
+    if (ARG_OUT_TARGET_NAME)
+        set(${ARG_OUT_TARGET_NAME} ${LOCAL_TARGET_NAME} PARENT_SCOPE)
+    endif()
 
 endfunction(enable_clang_format_check_for_directory)
-
-#
-# :function: enable_clang_format_in_place_for_directory
-# Create a target that reformats source, in-place, based on formatting rules.
-#
-# :param DIRECTORY path       - If provided the directory otherwise this is
-#                               ${CMAKE_CURRENT_SOURCE_DIR}
-# :param GLOB_PATTERN glob    - A pattern to match files against.
-#
-function(enable_clang_format_in_place_for_directory)
-
-    #+-[input]----------------------------------------------------------------+
-    set(options "")
-    set(singleValueArgs GLOB_PATTERN DIRECTORY)
-    set(multiValueArgs "")
-    cmake_parse_arguments(PARSE_ARGV 0 ARG "${options}" "${singleValueArgs}" "${multiValueArgs}")
-
-    if(NOT ARG_DIRECTORY)
-        set(ARG_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
-    endif()
-    #+-[body]-----------------------------------------------------------------+
-    cmake_path(GET CMAKE_CURRENT_SOURCE_DIR STEM LOCAL_DIRECTORY_NAME)
-    set(LOCAL_TARGET_NAME "danger-danger-${LOCAL_DIRECTORY_NAME}-clang-format-in-place")
-
-    add_custom_target(${LOCAL_TARGET_NAME}
-                      COMMAND ${CLANG_FORMAT_PYTHON_SHIM}
-                              --clang-format-path ${CLANG_FORMAT}
-                              -i
-                              ${ARG_GLOB_PATTERN}
-                      VERBATIM
-                      WORKING_DIRECTORY ${ARG_DIRECTORY}
-    )
-endfunction(enable_clang_format_in_place_for_directory)
