+++ cetlvast/suites/docs/footer.html
@@ -6,7 +6,7 @@
     $navpath
     <li class="footer"><address class="footer"><small>
       <a href="https://opencyphal.org">OpenCyphal</a>
-      – Copyright (C) 2014-2023 Pavel Kirienko <pavel.kirienko@gmail.com>
+      – Copyright (C) 2014-2023 Pavel Kirienko <pavel@opencyphal.org>
       – Copyright Amazon.com Inc. or its affiliates.<br/>This documentation is provided under the MIT license as part of the <a href="https://github.com/OpenCyphal/CETL/blob/master/LICENSE">CETL project</a>.
     </small></address></li>
   </ul>
@@ -15,7 +15,7 @@
 <!--BEGIN !GENERATE_TREEVIEW-->
 <hr class="footer"/><address class="footer"><small>
   <a href="https://opencyphal.org">OpenCyphal</a>
-  – Copyright (C) 2014-2023 Pavel Kirienko <pavel.kirienko@gmail.com>
+  – Copyright (C) 2014-2023 Pavel Kirienko <pavel@opencyphal.org>
   – Copyright Amazon.com Inc. or its affiliates.<br/>This documentation is provided under the MIT license as part of the <a href="https://github.com/OpenCyphal/CETL/blob/master/LICENSE">CETL project</a>.
 </small></address>
 <!--END !GENERATE_TREEVIEW-->
