+++ cetlvast/suites/unittest/sonar.cpp
@@ -1,0 +1,15 @@
+/// @file
+/// Hack to get sonarqube to analyze cetl headers.
+///
+/// @copyright
+/// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
+/// Copyright Amazon.com Inc. or its affiliates.
+/// SPDX-License-Identifier: MIT
+///
+
+#include "cetl/cetl.hpp"
+
+int main()
+{
+    return 0;
+}
